fix(eventstore): Prevent ENOENT error in setup
When installing the library from scratch, the test path / directory
used for storing the temporary event logs doesn't exist, which can
cause errors when attempting to truncate it.

Mitigated by adding a check as to whether the path exists, as well
as wrapping the `truncate` call in a `try...catch` block

+++ packages/eventstore/src/EventStore.test.ts
@@ -26,7 +26,14 @@ const stores = [
     opts: FILE_STORE_OPTS,
     setup: async () => {
       const filePath = path.resolve(FILE_STORE_OPTS.filepath);
-      await fs.truncate(filePath);
+      if (fs.exists(filePath)) {
+        try {
+          await fs.truncate(filePath);
+        } catch (e) {
+          console.warn('Error encountered during event log truncation:');
+          console.warn(e);
+        }
+      }
     },
     store: FileEventStore,
     type: 'FileEventStore'
